Repetir orden anterior I

+++ Drs.Model/Order/ItemModel.cs
@@ -2,7 +2,7 @@
 {
     public class ItemModel
     {
-        public int ItemId { get; set; }
+        public long ItemId { get; set; }
         public string Name { get; set; }
         public bool IsIdSpecified { get; set; }
         public double Price { get; set; }
